syslog-ng-common: remove P: Parser bound from Parser and ParserBuilder
This change simplifies the Parser and ParserBuilder traits. Note, that
the original approach forced the compiler to statically dispatch Pipe's
methods and the new approach forces dynamic dispatching.

This change has a little performance cost, but Pipe's methods are rarely
called and it's not worth to sacrifice simplicity for one use-case.

Signed-off-by: Tibor Benke <tibor.benke@balabit.com>